+++ src/App.test.jsx
@@ -1,10 +1,20 @@
 import React from 'react';
 
+import { useSelector } from 'react-redux';
+
 import { render } from '@testing-library/react';
 
 import App from './App';
 
+import regions from '../fixtures/regions';
+
+jest.mock('react-redux');
+
 describe('App', () => {
+  useSelector.mockImplementation((selector) => selector({
+    regions,
+  }))
+
   it('renders category buttons', () => {
     const { getByText } = render(<App />)
